Resolved findbug warnings

+++ src/Coalesce.Framework.Persister.Elasticsearch/src/main/java/com/incadencecorp/coalesce/framework/persistance/elasticsearch/ElasticSearchPersistor.java
@@ -228,7 +228,7 @@ public class ElasticSearchPersistor extends ElasticSearchTemplatePersister imple
                 }
 
                 // Back off to allow other threads to close their locks.
-                Thread.sleep(millis * attempt);
+                Thread.sleep((long) millis * attempt);
 
                 if (LOGGER.isDebugEnabled())
                 {
